Revert renames in client
