+++ .gitattributes
@@ -1,1 +1,1 @@
-client/docs/swagger-ui/* linguist-vendored
+client/lcd/swagger-ui/* linguist-vendored
