+++ client/lcd/statik/init.go
@@ -1,0 +1,3 @@
+package statik
+
+//This just for fixing the error in importing empty github.com/cosmos/cosmos-sdk/client/lcd/statik
