+++ client/lcd/statik/statik.go
