+++ client/lcd/swagger-ui/favicon-16x16.png
